stock 1.8.10 checkin

+++ api/classes/class.caction.php
@@ -27,10 +27,7 @@
  * Class containing methods for operations with Actions
  *
  */
-require_once 'include/actions.inc.php';
-
 class CAction extends CZBXAPI{
-
 /**
  * Get Actions data
  *
@@ -76,19 +73,20 @@ class CAction extends CZBXAPI{
 			'mediatypeids'			=> null,
 			'userids'				=> null,
 			'nopermissions'			=> null,
+			'editable'				=> null,
 // filter
-			'eventsource'			=> null,
-			'evaltype'				=> null,
-			'status'				=> null,
-			'esc_period'			=> null,
-			'recovery_msg'			=> null,
-			'pattern'				=> '',
+			'filter'				=> null,
+			'search'				=> null,
+			'startSearch'			=> null,
+			'excludeSearch'			=> null,
+			'searchWildcardsEnabled'=> null,
+
 // OutPut
 			'extendoutput'			=> null,
 			'output'				=> API_OUTPUT_REFER,
 			'select_conditions'		=> null,
 			'select_operations'		=> null,
-			'count'					=> null,
+			'countOutput'			=> null,
 			'preservekeys'			=> null,
 
 			'sortfield'				=> '',
@@ -320,46 +318,26 @@ class CAction extends CZBXAPI{
 // TODO:
 		}
 
-// eventsource
-		if(!is_null($options['eventsource'])){
-			$sql_parts['where'][] = 'a.eventsource='.$options['eventsource'];
+// filter
+		if(is_array($options['filter'])){
+			zbx_db_filter('actions a', $options, $sql_parts);
 		}
 
-// evaltype
-		if(!is_null($options['evaltype'])){
-			$sql_parts['where'][] = 'a.evaltype='.$options['evaltype'];
+// search
+		if(is_array($options['search'])){
+			zbx_db_search('actions a', $options, $sql_parts);
 		}
 
-// status
-		if(!is_null($options['status'])){
-			$sql_parts['where'][] = 'a.status='.$options['status'];
-		}
-
-// esc_period
-		if(!is_null($options['esc_period'])){
-			$sql_parts['where'][] = 'a.esc_period>'.$options['esc_period'];
-		}
-
-// recovery_msg
-		if(!is_null($options['recovery_msg'])){
-			$sql_parts['where'][] = 'a.recovery_msg<'.$options['recovery_msg'];
-		}
-
-// extendoutput
+// output
 		if($options['output'] == API_OUTPUT_EXTEND){
 			$sql_parts['select']['actions'] = 'a.*';
 		}
 
-// count
-		if(!is_null($options['count'])){
+// countOutput
+		if(!is_null($options['countOutput'])){
 			$options['sortfield'] = '';
 
 			$sql_parts['select'] = array('COUNT(DISTINCT a.actionid) as rowscount');
-		}
-
-// pattern
-		if(!zbx_empty($options['pattern'])){
-			$sql_parts['where'][] = ' UPPER(a.name) LIKE '.zbx_dbstr('%'.zbx_strtoupper($options['pattern']).'%');
 		}
 
 // order
@@ -407,8 +385,8 @@ class CAction extends CZBXAPI{
 		$db_res = DBselect($sql, $sql_limit);
 		while($action = DBfetch($db_res)){
 
-			if($options['count']){
-				$result = $action;
+			if($options['countOutput']){
+				$result = $action['rowscount'];
 			}
 			else{
 				$actionids[$action['actionid']] = $action['actionid'];
@@ -433,7 +411,7 @@ class CAction extends CZBXAPI{
 		}
 
 COpt::memoryPick();
-		if(($options['output'] != API_OUTPUT_EXTEND) || !is_null($options['count'])){
+		if(!is_null($options['countOutput'])){
 			if(is_null($options['preservekeys'])) $result = zbx_cleanHashes($result);
 			return $result;
 		}
@@ -490,6 +468,25 @@ COpt::memoryPick();
 	return $result;
 	}
 
+	public static function exists($object){
+		$keyFields = array(array('actionid', 'name'));
+
+		$options = array(
+			'filter' => zbx_array_mintersect($keyFields, $object),
+			'output' => API_OUTPUT_SHORTEN,
+			'nopermissions' => 1,
+			'limit' => 1
+		);
+
+		if(isset($object['node']))
+			$options['nodeids'] = getNodeIdByNodeName($object['node']);
+		else if(isset($object['nodeids']))
+			$options['nodeids'] = $object['nodeids'];
+
+		$objs = self::get($options);
+
+	return !empty($objs);
+	}
 /**
  * Add actions
  *
@@ -505,27 +502,43 @@ COpt::memoryPick();
  */
 	public static function create($actions){
 		$actions = zbx_toArray($actions);
-		$insert = array();
 		$conditions = array();
 		$operations = array();
 
 		try{
 			self::BeginTransaction(__METHOD__);
 
+// Check fields
+			$action_db_fields = array(
+				'name' => null,
+				'eventsource' => null,
+				'evaltype' => null,
+			);
+			$duplicates = array();
 			foreach($actions as $anum => $action){
-				$action_db_fields = array(
-					'name' => null,
-					'eventsource' => null,
-					'evaltype' => null,
-				);
 				if(!check_db_fields($action_db_fields, $action)){
-					self::exception(ZBX_API_ERROR_PARAMETERS, 'Incorrect parameters used for Action [ '.$action['name'].' ]');
+					self::exception(ZBX_API_ERROR_PARAMETERS, S_INCORRECT_PARAMETER_USED_FOR_ACTION.' [ '.$action['name'].' ]');
 				}
 
-				$insert[$anum] = $action;
+				if(isset($duplicates[$action['name']]))
+					self::exception(ZBX_API_ERROR_PARAMETERS, S_ACTION.' [ '.$action['name'].' ] '.S_ALREADY_EXISTS_SMALL);
+				else
+					$duplicates[$action['name']] = $action['name'];
 			}
 
-			$actionids = DB::insert('actions', $insert);
+			$options = array(
+				'filter' => array('name' => $duplicates),
+				'output' => API_OUTPUT_EXTEND,
+				'editable' => 1,
+				'nopermissions' => 1
+			);
+			$dbActions = self::get($options);
+			foreach($dbActions as $anum => $dbAction){
+				self::exception(ZBX_API_ERROR_PARAMETERS, S_ACTION.' [ '.$dbAction['name'].' ] '.S_ALREADY_EXISTS_SMALL);
+			}
+//------
+
+			$actionids = DB::insert('actions', $actions);
 
 			foreach($actions as $anum => $action){
 				if(isset($action['conditions']) && !empty($action['conditions'])){
@@ -535,7 +548,7 @@ COpt::memoryPick();
 				}
 
 				if(!isset($action['operations']) || empty($action['operations'])){
-					self::exception(ZBX_API_ERROR_PARAMETERS, 'Incorrect parameters used for Action [ '.$action['name'].' ]');
+					self::exception(ZBX_API_ERROR_PARAMETERS, S_INCORRECT_PARAMETER_USED_FOR_ACTION.' [ '.$action['name'].' ]');
 				}
 				else{
 					foreach($action['operations'] as $operation){
@@ -546,7 +559,6 @@ COpt::memoryPick();
 
 			self::addOperations($operations);
 			self::addConditions($conditions);
-
 
 			self::EndTransaction(true, __METHOD__);
 			return array('actionids' => $actionids);
@@ -574,13 +586,13 @@ COpt::memoryPick();
  * @param array $actions[0,...]['url'] OPTIONAL
  * @return boolean
  */
-	public static function  update($actions){
+	public static function update($actions){
 		$actions = zbx_toArray($actions);
 		$actionids = zbx_objectValues($actions, 'actionid');
+
 		$update = array();
 		$operations = array();
 		$conditions = array();
-
 		try{
 			self::BeginTransaction(__METHOD__);
 
@@ -597,13 +609,41 @@ COpt::memoryPick();
 				}
 			}
 
+// Check fields
+			$action_db_fields = array(
+				'actionid' => null
+			);
+
+			$duplicates = array();
 			foreach($actions as $anum => $action){
-				$action_db_fields = array(
-					'actionid' => null
-				);
 				if(!check_db_fields($action_db_fields, $action)){
-					self::exception(ZBX_API_ERROR_PARAMETERS, 'Incorrect parameters used for Action [ '.$action['name'].' ]');
+					self::exception(ZBX_API_ERROR_PARAMETERS, S_INCORRECT_PARAMETER_USED_FOR_ACTION.' [ '.$action['name'].' ]');
 				}
+
+				if(!isset($action['name'])) continue;
+
+				if(isset($duplicates[$action['name']]))
+					self::exception(ZBX_API_ERROR_PARAMETERS, S_ACTION.' [ '.$action['name'].' ] '.S_ALREADY_EXISTS_SMALL);
+				else
+					$duplicates[$action['name']] = $action['name'];
+			}
+//------
+
+			foreach($actions as $anum => $action){
+// Existance
+				if (isset($action['name'])) {
+					$options = array(
+						'filter' => array( 'name' => $action['name'] ),
+						'output' => API_OUTPUT_SHORTEN,
+						'editable' => 1,
+						'nopermissions' => 1
+					);
+					$action_exists = self::get($options);
+					if(($action_exist = reset($action_exists)) && ($action_exist['actionid'] != $action['actionid'])){
+						self::exception(ZBX_API_ERROR_PARAMETERS, S_ACTION.' [ '.$action['name'].' ] '.S_ALREADY_EXISTS_SMALL);
+					}
+				}
+//----
 
 				if(isset($action['conditions']) && !empty($action['conditions'])){
 					foreach($action['conditions'] as $condition){
@@ -611,11 +651,8 @@ COpt::memoryPick();
 					}
 				}
 
-				if(!isset($action['operations']) || empty($action['operations'])){
-					self::exception(ZBX_API_ERROR_PARAMETERS, 'Incorrect parameters used for Action [ '.$action['name'].' ]');
-				}
-				else{
-					foreach($action['operations'] as $operation){
+				if (isset($action['operations']) && !empty($action['operations'])) {
+					foreach ($action['operations'] as $operation) {
 						$operations[] = array_merge(array('actionid' => $action['actionid']), $operation);
 					}
 				}
@@ -632,20 +669,25 @@ COpt::memoryPick();
 
 			DB::update('actions', $update);
 
-			$operationids = array();
-			$sql = 'SELECT operationid FROM operations WHERE '.DBcondition('actionid', $actionids);
-			$operations_db = DBselect($sql);
-			while($operationid = DBfetch($operations_db)){
-				$operationids[] = $operationid['operationid'];
+			if (!empty($operations)) {
+				$operationids = array();
+				$sql = 'SELECT operationid FROM operations WHERE '.DBcondition('actionid', $actionids);
+				$operations_db = DBselect($sql);
+				while($operationid = DBfetch($operations_db)){
+					$operationids[] = $operationid['operationid'];
+				}
+
+				DB::delete('opconditions', DBcondition('operationid', $operationids));
+				DB::delete('opmediatypes', DBcondition('operationid', $operationids));
+				DB::delete('operations', DBcondition('actionid', $actionids));
+
+				self::addOperations($operations);
 			}
 
-			DB::delete('conditions', DBcondition('actionid', $actionids));
-			DB::delete('opconditions', DBcondition('operationid', $operationids));
-			DB::delete('opmediatypes', DBcondition('operationid', $operationids));
-			DB::delete('operations', DBcondition('actionid', $actionids));
-
-			self::addOperations($operations);
-			self::addConditions($conditions);
+			if (!empty($conditions)) {
+				DB::delete('conditions', DBcondition('actionid', $actionids));
+				self::addConditions($conditions);
+			}
 
 			self::EndTransaction(true, __METHOD__);
 			return array('actionids' => $actionids);
@@ -670,17 +712,16 @@ COpt::memoryPick();
  * @return boolean
  */
 	protected static function addConditions($conditions){
-      global $ZBX_MESSAGES;
-      $conditions = zbx_toArray($conditions);
+		$conditions = zbx_toArray($conditions);
 		$conditions_insert = array();
 
 		if(!check_permission_for_action_conditions($conditions)){
-			self::exception(ZBX_API_ERROR_PERMISSIONS, S_NOPERMISSIONS);
+			self::exception(ZBX_API_ERROR_PERMISSIONS, S_NO_PERMISSIONS);
 		}
 
 		foreach($conditions as $cnum => $condition){
 			if(!validate_condition($condition['conditiontype'], $condition['value'])){
-				self::exception(ZBX_API_ERROR_PARAMETERS, 'Incorrect parameters used for Conditions: '.print_r($condition));
+				self::exception(ZBX_API_ERROR_PARAMETERS, S_INCORRECT_PARAMETER_USED_FOR_CONDITIONS);
 			}
 			$conditions_insert[] = $condition;
 		}
@@ -693,7 +734,7 @@ COpt::memoryPick();
 /**
  * add operations
  *
- * @param _array $operations multidimensional array with operations data
+ * @param array $operations multidimensional array with operations data
  * @param array $operations[0,...]['actionid']
  * @param array $operations[0,...]['operationtype']
  * @param array $operations[0,...]['object']
@@ -730,7 +771,7 @@ COpt::memoryPick();
 				'operationtype' => null,
 			);
 			if(!check_db_fields($operation_db_fields, $operation)){
-				self::exception(ZBX_API_ERROR_PARAMETERS, 'Incorrect parameters used for Operations');
+				self::exception(ZBX_API_ERROR_PARAMETERS, S_INCORRECT_PARAMETER_USED_FOR_OPERATIONS);
 			}
 
 			$operation_inserts[$onum] = $operation;
